remove unnecessary changes

+++ lib/src/jsMain/kotlin/ch/ubique/libs/ktor/common/PathExtensions.js.kt
@@ -6,5 +6,5 @@ import kotlinx.io.files.Path
  * Kotlin/JS not implemented
  */
 internal actual fun Path.freeSpace(): Long {
-	return 0.toLong()
+	return 0
 }
